Move FeedImageViewModel to separate file
