+++ EssentialFeed/EssentialFeed/FeedPresentation/FeedImagePresenter.swift
@@ -6,18 +6,6 @@
 
 import Foundation
 
-public struct FeedImageViewModel<Image> {
-    public let description: String?
-    public let location: String?
-    public let image: Image?
-    public let isLoading: Bool
-    public let shouldRetry: Bool
-    
-    public var hasLocation: Bool {
-        return location != nil
-    }
-}
-
 public protocol FeedImageView {
     associatedtype Image
