+++ EssentialFeed/EssentialFeed/FeedPresentation/FeedImageViewModel.swift
@@ -1,0 +1,19 @@
+//
+//  FeedImageViewModel.swift
+//  EssentialFeed
+//
+//
+
+import Foundation
+
+public struct FeedImageViewModel<Image> {
+    public let description: String?
+    public let location: String?
+    public let image: Image?
+    public let isLoading: Bool
+    public let shouldRetry: Bool
+    
+    public var hasLocation: Bool {
+        return location != nil
+    }
+}
